backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
PyMuPDF>=1.25
pillow>=11.0
PyPDF2>=3.0
python-multipart>=0.0.20
arq>=0.26
redis[hiredis]>=5.0
pydantic-settings>=2.0
aiofiles>=24.0
aioboto3>=13.0
